+++ internal/domain/network/loadbalancer/load_balancer.go
@@ -37,6 +37,11 @@ type Backend struct {
 }
 
 type LoadBalancerRepository interface {
+	// Basic getters
 	GetLoadBalancer(ctx context.Context, ocid string) (*LoadBalancer, error)
 	ListLoadBalancers(ctx context.Context, compartmentID string) ([]LoadBalancer, error)
+
+	// Enriched getters (may perform additional API calls)
+	GetEnrichedLoadBalancer(ctx context.Context, ocid string) (*LoadBalancer, error)
+	ListEnrichedLoadBalancers(ctx context.Context, compartmentID string) ([]LoadBalancer, error)
 }
